requirements.txt
# Core ML and NLP
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
bertopic>=0.15.0
umap-learn>=0.5.3
hdbscan>=0.8.29

# OpenAI
openai>=1.0.0

# LlamaParse
llama-parse>=0.4.0

# Vector Database
chromadb>=0.4.0

# Knowledge Graph
neo4j>=5.12.0
py2neo>=2021.2.3

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Email Processing
email-validator>=2.0.0
python-dateutil>=2.8.0

# Content Generation
python-pptx>=0.6.21
Pillow>=10.0.0
reportlab>=4.0.0

# Video Generation
moviepy>=1.0.3
gtts>=2.3.0
pydub>=0.25.1

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
python-dotenv>=1.0.0
colorama>=0.4.6

# API and Web
flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
gunicorn>=21.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
